requirements.txt
python-dotenv>=0.21
supabase>=1.0.0
google-generativeai>=0.3.0
